Add credalbnn and clean benchmark code
