+++ src/probly_benchmark/method_scripts/credal_bnn.py
@@ -1,4 +1,4 @@
-"""Benchmarking for the credal relative likelihood method."""
+"""Benchmarking for the credal bnn method."""
 
 from __future__ import annotations
 
@@ -6,7 +6,7 @@ import logging
 
 import torch
 
-from probly.method.credal_relative_likelihood import credal_relative_likelihood
+from probly.method.credal_bnn import credal_bnn
 from probly.representer import representer
 from probly_benchmark.models import LeNet
 
@@ -14,9 +14,10 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 model = LeNet(n_classes=5)
-cep = credal_relative_likelihood(model, num_members=10)
+cep = credal_bnn(model, num_members=10)
 rep = representer(cep)
 logger.info(rep)
 inputs = torch.randn(3, 1, 28, 28)
 output = rep.predict(inputs)
 logger.info(output)
+logger.info(output.shape)
