+++ src/probly_benchmark/method_scripts/credal_ensembling.py
@@ -20,3 +20,4 @@ logger.info(rep)
 inputs = torch.randn(3, 1, 28, 28)
 output = rep.predict(inputs)
 logger.info(output)
+logger.info(output.shape)
