+++ src/probly_benchmark/method_scripts/credal_relative_likelihood.py
@@ -1,4 +1,4 @@
-"""DropConnect benchmark code."""
+"""Benchmarking for the credal relative likelihood method."""
 
 from __future__ import annotations
 
@@ -6,7 +6,7 @@ import logging
 
 import torch
 
-from probly.method.dropconnect import dropconnect
+from probly.method.credal_relative_likelihood import credal_relative_likelihood
 from probly.representer import representer
 from probly_benchmark.models import LeNet
 
@@ -14,8 +14,8 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 model = LeNet(n_classes=5)
-cep = dropconnect(model, p=0.5)
-rep = representer(cep, num_samples=10)
+cep = credal_relative_likelihood(model, num_members=10)
+rep = representer(cep)
 logger.info(rep)
 inputs = torch.randn(3, 1, 28, 28)
 output = rep.predict(inputs)
